Unify ArchiveReader and ArchiveWriter into Archive<M>
Replaces separate ArchiveReader and ArchiveWriter structs with a
generic Archive<M> that works with both MappedArchive (read-only) and
MappedArchiveMut (read-write) memory map types.

Introduces ArchiveRead and ArchiveWrite traits:
- ArchiveRead: entry_count, path_size, alignment, iter_entries,
  find_entry, read_data, verify_crc, is_sorted, find_duplicates,
  has_orphaned_data
- ArchiveWrite: add_entry, add_file, delete, sync (extends ArchiveRead)

Type aliases maintain backwards compatibility:
- ArchiveReader = Archive<MappedArchive>
- ArchiveWriter = Archive<MappedArchiveMut>

This unification prepares the codebase for FUSE support where both
read and write operations need access to common archive state.

Closes #42
